move invocation from main to run
