+++ main.go
@@ -18,59 +18,16 @@ package main
 
 import (
 	"os"
-	"os/signal"
-	"syscall"
 
-	"github.com/kardianos/service"
-	"github.com/netflix/weep/cmd"
+	"github.com/netflix/weep/run"
 	log "github.com/sirupsen/logrus"
 )
-
-var svcLogger service.Logger
-var done chan int
 
 func init() {
 	// Output to stdout instead of the default stderr
 	log.SetOutput(os.Stdout)
 }
 
-type program struct{}
-
-func (p *program) Start(s service.Service) error {
-	go p.run()
-	return nil
-}
-
-func (p *program) run() {
-	shutdown := make(chan os.Signal, 1)
-	done = make(chan int, 1)
-	signal.Notify(shutdown, syscall.SIGINT, syscall.SIGTERM)
-	cmd.Execute(shutdown, done)
-}
-
-func (p *program) Stop(s service.Service) error {
-	<-done
-	return nil
-}
-
 func main() {
-	svcConfig := &service.Config{
-		Name:        "Weep",
-		DisplayName: "Weep",
-		Description: "The ConsoleMe CLI",
-	}
-
-	prg := &program{}
-	s, err := service.New(prg, svcConfig)
-	if err != nil {
-		log.Fatal(err)
-	}
-	svcLogger, err = s.Logger(nil)
-	if err != nil {
-		log.Fatal(err)
-	}
-	err = s.Run()
-	if err != nil {
-		_ = svcLogger.Error(err)
-	}
+	run.Run()
 }
